Created class with tests for login to use in AUTH DB and AUTH LDAP
